refactor(frontend): Add lazy-loaded edit-assignment module
