+++ frontend/src/app/assignment/assignment-routing.module.ts
@@ -9,42 +9,11 @@ import {AssignmentTasksComponent} from './pages/assignment/tasks/tasks.component
 import {CourseComponent} from './pages/course/course.component';
 import {CreateCourseComponent} from './pages/create-course/create-course.component';
 import {CreateSolutionComponent} from './pages/create-solution/create-solution.component';
-import {ClassroomComponent} from './pages/edit-assignment/classroom/classroom.component';
-import {EditAssignmentComponent} from './pages/edit-assignment/edit-assignment.component';
-import {InfoComponent} from './pages/edit-assignment/info/info.component';
-import {PreviewComponent} from './pages/edit-assignment/preview/preview.component';
-import {SampleComponent} from './pages/edit-assignment/sample/sample.component';
-import {TasksComponent} from './pages/edit-assignment/tasks/tasks.component';
-import {TemplateComponent} from './pages/edit-assignment/template/template.component';
-import {EditTaskModalComponent} from './pages/edit-task-modal/edit-task-modal.component';
 import {MyAssignmentsComponent} from './pages/my-assignments/my-assignments.component';
 import {MyCoursesComponent} from './pages/my-courses/my-courses.component';
 import {MySolutionsComponent} from './pages/my-solutions/my-solutions.component';
 import {TokenModalComponent} from './pages/token-modal/token-modal.component';
 
-export const editAssignmentChildRoutes: Routes = [
-  {path: 'info', component: InfoComponent, data: {title: 'Info'}},
-  {path: 'classroom', component: ClassroomComponent, data: {icon: 'github', title: 'Classroom', new: true}},
-  {
-    path: 'tasks',
-    component: TasksComponent,
-    data: {title: 'Tasks'},
-    children: [
-      {path: ':task', component: EditTaskModalComponent},
-    ],
-  },
-  {path: 'template', component: TemplateComponent, data: {title: 'Template'}},
-  {
-    path: 'sample',
-    component: SampleComponent,
-    data: {title: 'Sample'},
-    children: [
-      {path: ':task', component: EditTaskModalComponent},
-    ],
-  },
-  {path: 'preview', component: PreviewComponent, data: {title: 'Preview'}},
-  {path: '', redirectTo: 'info'},
-];
 
 export const assignmentChildRoutes = [
   {path: 'tasks', component: AssignmentTasksComponent, data: {title: 'Tasks & Sample Solution'}},
@@ -64,8 +33,7 @@ const routes: Routes = [
   {path: 'courses/:cid/assignments/:aid', component: CourseComponent},
   {
     path: 'create',
-    component: EditAssignmentComponent,
-    children: editAssignmentChildRoutes,
+    loadChildren: () => import('./modules/edit-assignment/edit-assignment.module').then(m => m.EditAssignmentModule),
   },
   {path: ':aid/solutions/new', component: CreateSolutionComponent},
   {
@@ -74,8 +42,7 @@ const routes: Routes = [
   },
   {
     path: ':aid/edit',
-    component: EditAssignmentComponent,
-    children: editAssignmentChildRoutes,
+    loadChildren: () => import('./modules/edit-assignment/edit-assignment.module').then(m => m.EditAssignmentModule),
   },
   {path: ':aid', pathMatch: 'full', redirectTo: ':aid/solutions/new'},
   {
