+++ frontend/src/app/assignment/assignment.module.ts
@@ -12,9 +12,6 @@ import {DndModule} from 'ngx-drag-drop';
 import {SharedModule} from '../shared/shared.module';
 import {AssignmentRoutingModule} from './assignment-routing.module';
 
-import {AssignmentInfoComponent} from './components/assignment-info/assignment-info.component';
-import {EditTaskListComponent} from './components/edit-task-list/edit-task-list.component';
-import {ImportExportComponent} from './components/import-export/import-export.component';
 import {AssignmentSharedModule} from './modules/shared/shared.module';
 
 import {AssignmentComponent} from './pages/assignment/assignment.component';
@@ -25,14 +22,6 @@ import {AssignmentTasksComponent} from './pages/assignment/tasks/tasks.component
 import {CourseComponent} from './pages/course/course.component';
 import {CreateCourseComponent} from './pages/create-course/create-course.component';
 import {CreateSolutionComponent} from './pages/create-solution/create-solution.component';
-import {ClassroomComponent} from './pages/edit-assignment/classroom/classroom.component';
-import {EditAssignmentComponent} from './pages/edit-assignment/edit-assignment.component';
-import {InfoComponent} from './pages/edit-assignment/info/info.component';
-import {PreviewComponent} from './pages/edit-assignment/preview/preview.component';
-import {SampleComponent} from './pages/edit-assignment/sample/sample.component';
-import {TasksComponent} from './pages/edit-assignment/tasks/tasks.component';
-import {TemplateComponent} from './pages/edit-assignment/template/template.component';
-import {EditTaskModalComponent} from './pages/edit-task-modal/edit-task-modal.component';
 import {MyAssignmentsComponent} from './pages/my-assignments/my-assignments.component';
 import {MyCoursesComponent} from './pages/my-courses/my-courses.component';
 import {MySolutionsComponent} from './pages/my-solutions/my-solutions.component';
@@ -41,27 +30,16 @@ import {TokenModalComponent} from './pages/token-modal/token-modal.component';
 @NgModule({
   declarations: [
     AssignmentComponent,
-    AssignmentInfoComponent,
     AssignmentTasksComponent,
-    ClassroomComponent,
     CourseComponent,
     CreateCourseComponent,
     CreateSolutionComponent,
-    EditAssignmentComponent,
-    EditTaskListComponent,
-    EditTaskModalComponent,
-    ImportExportComponent,
-    InfoComponent,
     MyAssignmentsComponent,
     MyCoursesComponent,
     MySolutionsComponent,
-    PreviewComponent,
-    SampleComponent,
     ShareComponent,
     SolutionTableComponent,
     StatisticsComponent,
-    TasksComponent,
-    TemplateComponent,
     TokenModalComponent,
   ],
   imports: [
@@ -72,7 +50,6 @@ import {TokenModalComponent} from './pages/token-modal/token-modal.component';
     NgbModule,
     ClipboardModule,
     DragulaModule,
-    DndModule,
     AssignmentSharedModule,
     AssignmentRoutingModule,
     BarChartModule,
