+++ frontend/src/app/assignment/modules/edit-assignment/info/info.component.html
