+++ frontend/src/app/assignment/modules/shared/assignment-info/assignment-info.component.html
