+++ frontend/src/app/assignment/modules/shared/shared.module.ts
@@ -2,7 +2,10 @@ import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {RouterModule} from '@angular/router';
 import {NgbTooltipModule} from '@ng-bootstrap/ng-bootstrap';
+import {SharedModule} from '../../../shared/shared.module';
+import {AssignmentInfoComponent} from './assignment-info/assignment-info.component';
 import {AuthorNameComponent} from './author-name/author-name.component';
+import {ImportExportComponent} from './import-export/import-export.component';
 import {TaskColorPipe} from './pipes/task-color.pipe';
 import {TaskListComponent} from './task-list/task-list.component';
 
@@ -12,14 +15,19 @@ import {TaskListComponent} from './task-list/task-list.component';
     CommonModule,
     RouterModule,
     NgbTooltipModule,
+    SharedModule,
   ],
   declarations: [
+    AssignmentInfoComponent,
     AuthorNameComponent,
+    ImportExportComponent,
     TaskListComponent,
     TaskColorPipe,
   ],
   exports: [
+    AssignmentInfoComponent,
     AuthorNameComponent,
+    ImportExportComponent,
     TaskListComponent,
     TaskColorPipe,
   ],
